requirements.txt
PyPDF2>=3.0.0
pdfminer.six>=20221105
pdfplumber>=0.9.0
aiohttp>=3.8.0
pydantic>=1.10.0
pydantic-settings>=1.3.0
openai>=0.27.8
pytesseract>=0.3.10
tiktoken>=0.5.0
langchain-openai>=0.2.0
markdown2>=2.4.10
pdfkit>=1.0.0
markdown-pdf>=0.11.0
minio>=7.1.8
pymilvus>=1.2.3
sqlalchemy>=2.0.21
fastapi>=0.103.0
alembic>=1.12.0
asyncpg>=0.29.0
greenlet>=2.1.4
uvicorn==0.32.1